Newest vers. data cruncher backend
Input: An array with the IR sensor's current angle (degrees) and the IR reading from a SHARP IR when scanning an object.
Also, a calibrated array with the same readings but no object being scanned.

Output: Currently graphs of the calibration line of best fit, the angle of the IR vs the IR readings, and the angle of the IR vs. the unadjusted distance in cm.